+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/input/MiniAppByUrlActivity.kt
@@ -1,13 +1,18 @@
 package com.rakuten.tech.mobile.testapp.ui.input
 
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
+import android.util.Patterns
 import android.webkit.URLUtil
 import androidx.databinding.DataBindingUtil
 import com.rakuten.tech.mobile.miniapp.testapp.R
 import com.rakuten.tech.mobile.miniapp.testapp.databinding.MiniAppInputActivityBinding
+import com.rakuten.tech.mobile.testapp.helper.isValidMiniAppUrl
 import com.rakuten.tech.mobile.testapp.ui.base.BaseActivity
+import com.rakuten.tech.mobile.testapp.ui.deeplink.SchemeByUrlActivity
 import com.rakuten.tech.mobile.testapp.ui.display.MiniAppDisplayActivity
 import com.rakuten.tech.mobile.testapp.ui.display.preload.PreloadMiniAppWindow
 
@@ -20,10 +25,18 @@ class MiniAppByUrlActivity : BaseActivity(), PreloadMiniAppWindow.PreloadMiniApp
 
     sealed class InputDisplay(val input: String) {
         class Url(input: String) : InputDisplay(input)
+        class MiniAppUrl(input: String) : InputDisplay(input)
         class None : InputDisplay("")
     }
 
     private var display: InputDisplay = InputDisplay.None()
+    private var miniappCameraPermissionCallback: (isGranted: Boolean) -> Unit = {
+        if (it) {
+            val intent = Intent(Intent.ACTION_VIEW, Uri.parse(display.input.trim()))
+            startActivity(intent)
+        }
+    }
+
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -58,18 +71,28 @@ class MiniAppByUrlActivity : BaseActivity(), PreloadMiniAppWindow.PreloadMiniApp
         finish()
     }
 
+    private fun isValidAppUrl(input: String): Boolean = URLUtil.isValidUrl(input)
+            && Patterns.WEB_URL.matcher(input).matches()
 
-    private fun validateInput(input: String) {
+    fun validateInput(input: String) {
         if (input.isBlank())
             binding.btnDisplayAppId.isEnabled = false
         else {
-            display = if (URLUtil.isValidUrl(input)) {
-                onValidUI()
-                InputDisplay.Url(input)
-            } else {
-                onInvalidUI()
-                InputDisplay.None()
-            }
+            display =
+                when {
+                    (isValidAppUrl(input)) -> {
+                        onValidUI()
+                        InputDisplay.Url(input)
+                    }
+                    isValidMiniAppUrl(input) -> {
+                        onValidUI()
+                        InputDisplay.MiniAppUrl(input)
+                    }
+                    else -> {
+                        onInvalidUI()
+                        InputDisplay.None()
+                    }
+                }
         }
     }
 
@@ -84,7 +107,18 @@ class MiniAppByUrlActivity : BaseActivity(), PreloadMiniAppWindow.PreloadMiniApp
     }
 
     private fun displayMiniApp() = when (display) {
-        is InputDisplay.Url -> MiniAppDisplayActivity.startUrl(this, display.input.trim())
+        is InputDisplay.Url -> {
+            MiniAppDisplayActivity.startUrl(
+                this,
+                display.input.trim(),
+            )
+        }
+        is InputDisplay.MiniAppUrl -> {
+            SchemeByUrlActivity.start(
+                this,
+                display.input.trim()
+            )
+        }
         is InputDisplay.None -> {
             //intended
         }